minute chnges

+++ Hackerrank/Code Actually/day8.cpp
@@ -12,9 +12,9 @@ int main(){
         string A, B;
         cin>>A>>B;
         bool flag = true;
-        int dif = (B[0] - A[0]+26)%26;
-        for(long long int i=1;i<A.length();i++)
-            if((B[i] - A[i]+26)%26!=dif )
+        int dif = (B[0] - A[0] + 26) % 26;
+        for(long int i=1; i<A.length(); i++)
+            if((B[i] - A[i] + 26 ) %26 !=dif )
                 flag = false;
         if(flag)
             cout<<dif<<"\n";
